+++ src/array.c
@@ -7,7 +7,7 @@ void generic_array_free(struct generic_array *a) {
     a->length = 0;
 }
 
-int generic_array_ensure_size(struct generic_array *a, size_t element_size, int requested_size) {
+int generic_array_ensure_size(struct generic_array *a, int element_size, int requested_size) {
     if (a->size >= requested_size) return 0;
     int new_size = (a->size < 8 ? 8 : a->size);
     while (new_size < requested_size) {
